requirements.txt
# OpenVINO GenAI (for NPU/GPU/CPU)
openvino-genai>=2025.0.0
openvino>=2025.0.0

# Core dependencies
numpy>=1.26.0
huggingface-hub>=0.20.0
flask>=3.0.0
flask-restx>=1.3.0

# Embedding model support
optimum-intel>=1.26.0
transformers>=4.40.0
torch>=2.0.0
